changed array types to int

+++ Server/include/mergesort.h
@@ -1,12 +1,12 @@
 #pragma once
 #include<vector>
 
-void Merge( std::vector<double>& arr, std::vector<double>& copy, int l, int m, int r);
+void Merge( std::vector<int>& arr, std::vector<int>& copy, int l, int m, int r);
 
-void MergeSortRec(std::vector<double>& arr, std::vector<double>& copy, int l, int r);
+void MergeSortRec(std::vector<int>& arr, std::vector<int>& copy, int l, int r);
 
-void MergeSort(std::vector<double>& arr);
+void MergeSort(std::vector<int>& arr);
 
-void ParallelMergeSortRec(std::vector<double>& arr, std::vector<double>& copy, int l, int r, int depth);
+void ParallelMergeSortRec(std::vector<int>& arr, std::vector<int>& copy, int l, int r, int depth);
 
-void ParallelMergeSort(std::vector<double>& arr, size_t n_threads);
+void ParallelMergeSort(std::vector<int>& arr, size_t n_threads);
